Use SourceGenerator for native code of PixelCollection.

+++ src/Magick.NET/Netstandard21/Pixels/PixelCollection.cs
@@ -44,7 +44,7 @@ internal abstract partial class PixelCollection
         => GetReadOnlyArea(geometry.X, geometry.Y, geometry.Width, geometry.Height);
 
     public virtual void SetArea(int x, int y, int width, int height, ReadOnlySpan<QuantumType> values)
-        => SetAreaUnchecked(x, y, width, height, values);
+        => SetAreaUnchecked(x, y, (uint)width, (uint)height, values);
 
     public virtual void SetArea(IMagickGeometry geometry, ReadOnlySpan<QuantumType> values)
         => SetArea(geometry.X, geometry.Y, geometry.Width, geometry.Height, values);
@@ -53,13 +53,13 @@ internal abstract partial class PixelCollection
         => SetPixelUnchecked(x, y, value);
 
     public virtual void SetPixels(ReadOnlySpan<QuantumType> values)
-        => SetAreaUnchecked(0, 0, Image.Width, Image.Height, values);
+        => SetAreaUnchecked(0, 0, (uint)Image.Width, (uint)Image.Height, values);
 
     private void SetPixelUnchecked(int x, int y, ReadOnlySpan<QuantumType> value)
         => SetAreaUnchecked(x, y, 1, 1, value);
 
-    private void SetAreaUnchecked(int x, int y, int width, int height, ReadOnlySpan<QuantumType> values)
-        => NativeInstance.SetArea(x, y, width, height, values, values.Length);
+    private void SetAreaUnchecked(int x, int y, uint width, uint height, ReadOnlySpan<QuantumType> values)
+        => NativeInstance.SetArea(x, y, width, height, values, (uint)values.Length);
 }
 
 #endif
