+++ src/Magick.NET/Pixels/PixelCollection.cs
@@ -48,13 +48,13 @@ internal abstract partial class PixelCollection : IPixelCollection<QuantumType>
     }
 
     public virtual QuantumType[]? GetArea(int x, int y, int width, int height)
-        => GetAreaUnchecked(x, y, width, height);
+        => GetAreaUnchecked(x, y, (uint)width, (uint)height);
 
     public virtual QuantumType[]? GetArea(IMagickGeometry geometry)
         => GetArea(geometry.X, geometry.Y, geometry.Width, geometry.Height);
 
     public virtual IntPtr GetAreaPointer(int x, int y, int width, int height)
-        => NativeInstance.GetArea(x, y, width, height);
+        => NativeInstance.GetArea(x, y, (uint)width, (uint)height);
 
     public virtual IntPtr GetAreaPointer(IMagickGeometry geometry)
         => GetAreaPointer(geometry.X, geometry.Y, geometry.Width, geometry.Height);
@@ -63,7 +63,7 @@ internal abstract partial class PixelCollection : IPixelCollection<QuantumType>
         => GetEnumerator();
 
     public IEnumerator<IPixel<QuantumType>> GetEnumerator()
-        => new PixelCollectionEnumerator(this, Image.Width, Image.Height);
+        => new PixelCollectionEnumerator(this, (uint)Image.Width, (uint)Image.Height);
 
     public int GetChannelIndex(PixelChannel channel)
         => Image.ChannelOffset(channel);
@@ -78,10 +78,10 @@ internal abstract partial class PixelCollection : IPixelCollection<QuantumType>
         => GetAreaUnchecked(x, y, 1, 1);
 
     public QuantumType[]? GetValues()
-        => GetAreaUnchecked(0, 0, Image.Width, Image.Height);
+        => GetAreaUnchecked(0, 0, (uint)Image.Width, (uint)Image.Height);
 
     public virtual void SetArea(int x, int y, int width, int height, QuantumType[] values)
-        => SetAreaUnchecked(x, y, width, height, values);
+        => SetAreaUnchecked(x, y, (uint)width, (uint)height, values);
 
     public virtual void SetArea(IMagickGeometry geometry, QuantumType[] values)
         => SetArea(geometry.X, geometry.Y, geometry.Width, geometry.Height, values);
@@ -89,7 +89,7 @@ internal abstract partial class PixelCollection : IPixelCollection<QuantumType>
     public virtual void SetByteArea(int x, int y, int width, int height, byte[] values)
     {
         var castedValues = CastArray(values, Quantum.Convert);
-        SetAreaUnchecked(x, y, width, height, castedValues);
+        SetAreaUnchecked(x, y, (uint)width, (uint)height, castedValues);
     }
 
     public virtual void SetByteArea(IMagickGeometry geometry, byte[] values)
@@ -98,13 +98,13 @@ internal abstract partial class PixelCollection : IPixelCollection<QuantumType>
     public virtual void SetBytePixels(byte[] values)
     {
         var castedValues = CastArray(values, Quantum.Convert);
-        SetAreaUnchecked(0, 0, Image.Width, Image.Height, castedValues);
+        SetAreaUnchecked(0, 0, (uint)Image.Width, (uint)Image.Height, castedValues);
     }
 
     public virtual void SetDoubleArea(int x, int y, int width, int height, double[] values)
     {
         var castedValues = CastArray(values, Quantum.Convert);
-        SetAreaUnchecked(x, y, width, height, castedValues);
+        SetAreaUnchecked(x, y, (uint)width, (uint)height, castedValues);
     }
 
     public virtual void SetDoubleArea(IMagickGeometry geometry, double[] values)
@@ -113,13 +113,13 @@ internal abstract partial class PixelCollection : IPixelCollection<QuantumType>
     public virtual void SetDoublePixels(double[] values)
     {
         var castedValues = CastArray(values, Quantum.Convert);
-        SetAreaUnchecked(0, 0, Image.Width, Image.Height, castedValues);
+        SetAreaUnchecked(0, 0, (uint)Image.Width, (uint)Image.Height, castedValues);
     }
 
     public virtual void SetIntArea(int x, int y, int width, int height, int[] values)
     {
         var castedValues = CastArray(values, Quantum.Convert);
-        SetAreaUnchecked(x, y, width, height, castedValues);
+        SetAreaUnchecked(x, y, (uint)width, (uint)height, castedValues);
     }
 
     public virtual void SetIntArea(IMagickGeometry geometry, int[] values)
@@ -128,7 +128,7 @@ internal abstract partial class PixelCollection : IPixelCollection<QuantumType>
     public virtual void SetIntPixels(int[] values)
     {
         var castedValues = CastArray(values, Quantum.Convert);
-        SetAreaUnchecked(0, 0, Image.Width, Image.Height, castedValues);
+        SetAreaUnchecked(0, 0, (uint)Image.Width, (uint)Image.Height, castedValues);
     }
 
     public virtual void SetPixel(int x, int y, QuantumType[] value)
@@ -151,7 +151,7 @@ internal abstract partial class PixelCollection : IPixelCollection<QuantumType>
     }
 
     public virtual void SetPixels(QuantumType[] values)
-        => SetAreaUnchecked(0, 0, Image.Width, Image.Height, values);
+        => SetAreaUnchecked(0, 0, (uint)Image.Width, (uint)Image.Height, values);
 
     public QuantumType[]? ToArray()
         => GetValues();
@@ -163,7 +163,7 @@ internal abstract partial class PixelCollection : IPixelCollection<QuantumType>
 
         try
         {
-            nativeResult = NativeInstance.ToByteArray(x, y, width, height, mapping);
+            nativeResult = NativeInstance.ToByteArray(x, y, (uint)width, (uint)height, mapping);
             result = ByteConverter.ToArray(nativeResult, width * height * mapping.Length);
         }
         finally
@@ -195,7 +195,7 @@ internal abstract partial class PixelCollection : IPixelCollection<QuantumType>
 
         try
         {
-            nativeResult = NativeInstance.ToShortArray(x, y, width, height, mapping);
+            nativeResult = NativeInstance.ToShortArray(x, y, (uint)width, (uint)height, mapping);
             return ShortConverter.ToArray(nativeResult, width * height * mapping.Length);
         }
         finally
@@ -219,13 +219,13 @@ internal abstract partial class PixelCollection : IPixelCollection<QuantumType>
     public ushort[]? ToShortArray(PixelMapping mapping)
         => ToShortArray(0, 0, Image.Width, Image.Height, mapping.ToString());
 
-    internal QuantumType[]? GetAreaUnchecked(int x, int y, int width, int height)
+    internal QuantumType[]? GetAreaUnchecked(int x, int y, uint width, uint height)
     {
         var pixels = NativeInstance.GetArea(x, y, width, height);
         if (pixels == IntPtr.Zero)
             throw new InvalidOperationException("Image contains no pixel data.");
 
-        var length = width * height * Image.ChannelCount;
+        var length = width * height * (uint)Image.ChannelCount;
         return QuantumConverter.ToArray(pixels, length);
     }
 
@@ -241,6 +241,6 @@ internal abstract partial class PixelCollection : IPixelCollection<QuantumType>
         return result;
     }
 
-    private void SetAreaUnchecked(int x, int y, int width, int height, QuantumType[] values)
-        => NativeInstance.SetArea(x, y, width, height, values, values.Length);
+    private void SetAreaUnchecked(int x, int y, uint width, uint height, QuantumType[] values)
+        => NativeInstance.SetArea(x, y, width, height, values, (uint)values.Length);
 }
